refactor(arch): reorganize directory structure to feature-driven layout
- Restructured directories into global (components) vs domain-specific (features) namespaces
- Moved home, about, contact, and events specific components to features/
- Moved routing files to pages/ and consolidated Index/LandingPage into home/HomePage.tsx
- Standardized hook file naming (camelCase) and file extensions
- Updated all internal paths and imports to use the @ path alias
