+++ src/App.tsx
@@ -1,12 +1,12 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Index from './pages/Index';
-import ContactPage from './pages/contact/ContactPage';
-import AboutPage from './components/AboutPage';
-import EventsPage from './pages/events/EventsPage';
-import Navbar from './components/layout/Navbar';
-import Footer from './components/layout/Footer';
-import SmoothScroll from './components/layout/SmoothScroll';
+import HomePage from '@/pages/home/HomePage';
+import ContactPage from '@/pages/contact/ContactPage';
+import AboutPage from '@/pages/about/AboutPage';
+import EventsPage from '@/pages/events/EventsPage';
+import Navbar from '@/components/layout/Navbar';
+import Footer from '@/components/layout/Footer';
+import SmoothScroll from '@/components/layout/SmoothScroll';
 
 const queryClient = new QueryClient();
 
@@ -16,7 +16,7 @@ const App = () => (
       {/* <SmoothScroll /> */}
       <Navbar />
       <Routes>
-        <Route path="/" element={<Index />} />
+        <Route path="/" element={<HomePage />} />
         <Route path="/about" element={<AboutPage />} />
         <Route path="/events" element={<EventsPage />} />
         <Route path="/contact" element={<ContactPage />} />
