--- src/components/LandingPage.tsx
@@ -1,19 +1,0 @@
-import HeroSection from './home/HeroSection';
-import AboutSection from './home/AboutSection';
-import Divisions from './home/DivisionsSection';
-import Hands3D from './home/Hands3D';
-import ApplyNowFormSection from './home/ApplyNowFormSection';
-
-const Landing = () => {
-  return (
-    <>
-      <HeroSection />
-      <AboutSection />
-      <Divisions />
-      <Hands3D />
-      <ApplyNowFormSection />
-    </>
-  );
-};
-
-export default Landing;
