+++ src/components/layout/Navbar.tsx
@@ -1,7 +1,7 @@
 import { useState, useEffect, useRef } from 'react';
 import { NavLink, Link } from 'react-router-dom';
 import { Menu, X } from 'lucide-react';
-import logo from '../../assets/seds-logo.png';
+import logo from '@/assets/seds-logo.png';
 
 const Navbar = () => {
   const [isOpen, setIsOpen] = useState(false);
