+++ src/features/about/components/AboutDivisionsSection.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import image1 from "../../assets/divisions/image1.png";
-import image2 from "../../assets/divisions/image2.png";
-import image3 from "../../assets/divisions/image3.png";
-import image4 from "../../assets/divisions/image4.png";
+import image1 from "@/assets/divisions/image1.png";
+import image2 from "@/assets/divisions/image2.png";
+import image3 from "@/assets/divisions/image3.png";
+import image4 from "@/assets/divisions/image4.png";
 
 interface Division {
   id: number;
