+++ src/features/contact/components/ContactForm.tsx
@@ -1,4 +1,4 @@
-import Button from '../ui/Button';
+import Button from '@/components/ui/Button';
 
 const ContactForm = () => {
   return (
