+++ src/features/home/components/AboutSection.tsx
