+++ src/features/home/components/Hands3D.tsx
