--- src/pages/Index.tsx
@@ -1,13 +1,0 @@
-import Landing from '../components/LandingPage';
-
-const Index = () => {
-  return (
-    <main className="relative min-h-screen">
-      <div className="relative z-10 flex flex-col w-full">
-        <Landing />
-      </div>
-    </main>
-  );
-};
-
-export default Index;
